seperated properties from pane into buffer display
