+++ src/buffer.rs
@@ -1,13 +1,9 @@
-use std::cmp::{self, max};
-use std::{borrow::Cow, cmp::min};
+use std::cmp::max;
+use std::cmp::min;
 
 use mlua::{prelude::LuaError, Lua, MetaMethod, ToLua, UserData, UserDataMethods};
 use ropey::Rope;
-use tree_sitter::{InputEdit, Language, Node, Parser, Point, Tree};
-use tui::{
-	style::{Color, Style},
-	text::Span,
-};
+use tree_sitter::{InputEdit, Language, Parser, Point, Tree};
 
 /// State that the buffer can undo to
 struct Revision<'a> {
@@ -27,30 +23,6 @@ pub struct Buffer {
 	pub tree: Option<Box<Tree>>,
 	// TODO: we need to support custom tab width rendering
 	tabwidth: u8,
-}
-
-// TODO: need to actually make a Theme dict lol
-/// Styles a string accordingly to the string and returns a Span
-///
-/// * text - The text that will be styled into a [`Span`]
-/// * token - The token name, expected to be returned from [`Node`].kind()
-fn write_token<'a, T>(text: T, token: &'static str) -> Span<'a>
-where
-	T: Into<Cow<'a, str>>,
-{
-	Span::styled(
-		text,
-		Style::default().fg(
-			// TODO: we really need to make the theme into a dict of some sort
-			match token {
-				"function" => Color::Rgb(246, 199, 255),
-				"identifier" => Color::Cyan,
-				"string" => Color::Yellow,
-				"comment" => Color::Green,
-				_ => Color::White,
-			},
-		),
-	)
 }
 
 impl Buffer {
@@ -90,93 +62,9 @@ impl Buffer {
 
 	/// A wrapper to converting a character index in the buffer content to
 	/// a point
-	fn point_to_char(&self, point: Point) -> usize {
+	pub fn point_to_char(&self, point: Point) -> usize {
 		let index = self.content.line_to_char(point.row);
 		index + point.column
-	}
-
-	/// Applies syntax highlighting to a line in the buffer
-	///
-	/// * line - The line of the content to highlight
-	/// * offset - The X offset of the view area to start rendering
-	/// * width - The width of the view area that will be rendered
-	pub fn highlight_line<'b>(&self, line: usize, offset: usize, width: usize) -> Vec<Span> {
-		let tree = self.tree.as_ref().unwrap();
-		let node = tree
-			.root_node()
-			.descendant_for_point_range(
-				Point {
-					column: 0,
-					row: line,
-				},
-				Point {
-					column: self.content.line(line).len_chars(),
-					row: line,
-				},
-			)
-			.unwrap();
-		let start = self.content.line_to_char(line) + offset;
-		let end = min(width + start, self.content.line_to_char(line + 1));
-		self.highlight(node, start, end)
-	}
-
-	/// Applies syntax highlighting to a region in the buffer
-	///
-	/// * node - The [`Node`] data that that spans the region of start and end.
-	/// * start - The start character point of the region
-	/// * end - The end character point of the region
-	///
-	/// # Example
-	///
-	/// ```rust
-	/// // this assumes that the buffer has the following content
-	/// // "function hello() { console.log('Hello, World!') }"
-	/// // let's grab the region of "function"
-	/// let start = 0;
-	/// let end = 8;
-	/// // get the node that spans the region of start and end
-	/// let node = buffer.tree.unwrap().root_node().descendent_for_byte_range(
-	///		start, end
-	/// );
-	/// let spans = Spans::from(buffer.highlight(node, start, end));
-	/// ```
-	pub fn highlight<'b>(&self, node: Node, start: usize, end: usize) -> Vec<Span> {
-		let cursor = &mut node.walk();
-		let mut vector: Vec<Span> = vec![];
-		let mut token_end = start;
-		loop {
-			// we select if it is a kind of "string" because the children of
-			// the "string" are the symbols surrounding the string and doesn't
-			// include the literal between them
-			if cursor.node().kind() == "string"
-				|| cursor.node().kind() == "comment"
-				|| !cursor.goto_first_child()
-			{
-				let start_char =
-					cmp::max(self.point_to_char(cursor.node().start_position()), start);
-				if start_char - token_end != 0 {
-					vector
-						.push(Span::raw(self.content.slice(
-							token_end.clamp(start, end)..start_char.clamp(start, end),
-						)));
-				}
-				vector.push(write_token(
-					self.content.slice(
-						start_char.clamp(start, end)
-							..self
-								.point_to_char(cursor.node().end_position())
-								.clamp(start, end),
-					),
-					cursor.node().kind(),
-				));
-				token_end = self.point_to_char(cursor.node().end_position());
-				while !cursor.goto_next_sibling() {
-					if !cursor.goto_parent() {
-						return vector;
-					}
-				}
-			}
-		}
 	}
 
 	/// Get an updated version of the parser tree
@@ -197,7 +85,7 @@ impl Buffer {
 
 	/// A wrapper to converting a character index in the buffer content to
 	/// a point
-	fn char_to_point(&self, char: usize) -> Point {
+	pub fn char_to_point(&self, char: usize) -> Point {
 		let row = self.content.char_to_line(char);
 		let column = char - self.content.line_to_char(row);
 		Point { row, column }
